Allow a DataBag list to be injected for nagios in chef solo
The NagiosDataBagHelper requires a DataBag at initialisation defaulting to a Chef::DataBag.list
call which bypasses the chef solo stuff, use the ability to inject a data bag and provide the
default list of nagios data bags and voila it works.

Also search for nodes on fqdn in solo mode and limit to those matching imos-* names so as not
to include dodgy node definitions that lack DNS entries as they can cause ip resolution breakage.