Allow an empty set of users for the user databag.
This allows us to only have pagerduty contacts. It also follows the
assumption in the contacts template where we do not add any user
contacts that do not have the nagios hash or nagios e-mail defined.

As in the template it ignores any users as contacts that do not
have the aforementioned items defined.

Change-Id: Iaf348015d9cca71720da5586707f00e937f1cf0d