Set server_name and server_alias in Apache2 to more sane defaults
The original defaults were based on early Chef ideas of how to setup a server (HJK probably).  Server name is best assumed to be the FQDN unless the user specifies a name they've set.  If they want a server alias it should be up to them to create this.  If not specified it shouldn't be set anyways.