Nagios on Debian 10 installation issue: dependencies package.
When installing Nagios on a debian 10 machine, the chef-client run fails because the recipe tries
to install the package libgd2-xpm-dev (among a few other errors).
By overriding node attributes, I was able to fix the installation, even if this cookbook doesn't seem
designed to work on debian 10. But this last problem was coming from a variable I could not override.

Details of the package is here https://packages.debian.org/fr/stretch/libgd2-xpm-dev.

In this PR, the dependencies are moved to default node attributes, so they can be overriden by cookbook's user.

Side-note:
I may come with a PR to fully support debian 10, but there is still work to do. This is not the purpose of this work.